fix(ios): resolve alerts without duplicate modal routing probes
